Merging with updates by francesco.
Server sends all the stations along a commute path as soon as the user
enters his home and work locations in CommuteActivity.
SearchOilRequest includes a new field 'km' to restrict the search to
a maximum distance from the user current position.

The list of oil stations is sent in multiple messages in order to
avoid error when the json message is too long to be received
correctly from the application's service.

I tested chat functionalities again and everything seems fine.